refactor: Consolidate and standardize user data constants across feed helpers
Improvements:
- Extract common user data constants to a centralized location
- Rename variables for clarity and consistency
- Add role-user mapping to improve seed generation
- Remove duplicate user data definitions in different feed helper files
